pyproject.toml
[tool.poetry]
name = "genai"
version = "0.1.0"
description = ""
authors = ["Samuel Davis <sam99dave@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.12"
langchain-core = "^0.2.28"
langchain-community = "^0.2.11"
langserve = {extras = ["all"], version = "^0.2.2"}
langchain-experimental = "^0.0.64"
python-fasthtml = "^0.2.1"
langchain-ollama = "^0.1.1"
redis = "^5.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
